PT-476 extracted base promotion evaluation policy class
